+++ src/http/types.rs
@@ -1,0 +1,121 @@
+use crate::types::Address;
+use serde::{Deserialize, Serialize};
+use std::collections::HashMap;
+use thiserror::Error;
+
+/// HTTP Request type that can be shared over WASM boundary to apps.
+#[derive(Debug, Serialize, Deserialize)]
+pub struct HttpRequest {
+    pub method: String,          // must parse to http::Method
+    pub version: Option<String>, // must parse to http::Version
+    pub url: String,             // must parse to url::Url
+    pub headers: HashMap<String, String>,
+    // BODY is stored in the payload, as bytes
+    // TIMEOUT is stored in the message expect_response
+}
+
+/// HTTP Response type that can be shared over WASM boundary to apps.
+#[derive(Debug, Serialize, Deserialize)]
+pub struct HttpResponse {
+    pub status: u16,
+    pub headers: HashMap<String, String>,
+    // BODY is stored in the payload, as bytes
+}
+
+#[derive(Error, Debug, Serialize, Deserialize)]
+pub enum HttpClientError {
+    #[error("http_client: request could not be parsed to HttpRequest: {}.", req)]
+    BadRequest { req: String },
+    #[error("http_client: http method not supported: {}", method)]
+    BadMethod { method: String },
+    #[error("http_client: url could not be parsed: {}", url)]
+    BadUrl { url: String },
+    #[error("http_client: http version not supported: {}", version)]
+    BadVersion { version: String },
+    #[error("http_client: failed to execute request {}", error)]
+    RequestFailed { error: String },
+}
+
+#[derive(Error, Debug, Serialize, Deserialize)]
+pub enum HttpServerError {
+    #[error("http_server: json is None")]
+    NoJson,
+    #[error("http_server: response not ok")]
+    ResponseError,
+    #[error("http_server: bytes are None")]
+    NoBytes,
+    #[error(
+        "http_server: JSON payload could not be parsed to HttpClientRequest: {error}. Got {:?}.",
+        json
+    )]
+    BadJson { json: String, error: String },
+    #[error("http_server: path binding error:  {:?}", error)]
+    PathBind { error: String },
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct JwtClaims {
+    pub username: String,
+    pub expiration: u64,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct WebSocketServerTarget {
+    pub node: String,
+    pub id: Option<String>,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct WebSocketPush {
+    pub target: WebSocketServerTarget,
+    pub is_text: Option<bool>,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct ServerAction {
+    pub action: String,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub enum HttpServerAction {
+    BindPath {
+        path: String,
+        authenticated: bool,
+        local_only: bool,
+    },
+    WebSocketPush(WebSocketPush),
+    ServerAction(ServerAction),
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct WsRegister {
+    pub ws_auth_token: String,
+    pub auth_token: String,
+    pub channel_id: String,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct WsMessage {
+    pub ws_auth_token: String,
+    pub auth_token: String,
+    pub channel_id: String,
+    pub target: Address,
+    pub json: Option<serde_json::Value>,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct EncryptedWsMessage {
+    pub ws_auth_token: String,
+    pub auth_token: String,
+    pub channel_id: String,
+    pub target: Address,
+    pub encrypted: String, // Encrypted JSON as hex with the 32-byte authentication tag appended
+    pub nonce: String,     // Hex of the 12-byte nonce
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub enum WebSocketClientMessage {
+    WsRegister(WsRegister),
+    WsMessage(WsMessage),
+    EncryptedWsMessage(EncryptedWsMessage),
+}
